+++ src/movegen.cpp
@@ -8,6 +8,6 @@ void genMoves(MoveList& moveList, const Board& board)
     while (moveLocations)
     {
         uint32_t sq = popLSB(moveLocations);
-        moveList.push_back(Move(sq));
+        moveList.push_back(Move(static_cast<uint16_t>(sq)));
     }
 }
